refactor(commands): extract startCmd registration from start.go
Moved the startCmd cobra.Command var and its init()/flag registration to
start_cmd.go (56 lines). Pure move, verified via go test. First step of the
711-line start.go split (T-P6-E2-W1-S1-T3).
